+++ keep-ui/middleware.tsx
@@ -2,20 +2,20 @@ import { NextResponse, type NextRequest } from "next/server";
 import { getToken } from "next-auth/jwt";
 import type { JWT } from "next-auth/jwt";
 import { getApiURL } from "@/utils/apiUrl";
+import { config as authConfig } from "@/auth.config";
+import NextAuth from "next-auth";
+
+const { auth } = NextAuth(authConfig);
 
 export async function middleware(request: NextRequest) {
   const { pathname, searchParams } = request.nextUrl;
 
+  const session = await auth();
+  const role = session?.userRole;
+  const isAuthenticated = !!session;
   // Keep it on header so it can be used in server components
   const requestHeaders = new Headers(request.headers);
   requestHeaders.set("x-url", request.url);
-
-  // Get the token using next-auth/jwt with the correct type
-  const token = (await getToken({
-    req: request,
-    secret: process.env.NEXTAUTH_SECRET,
-  })) as JWT | null;
-
   // Handle legacy /backend/ redirects
   if (pathname.startsWith("/backend/")) {
     const apiUrl = getApiURL();
@@ -33,13 +33,13 @@ export async function middleware(request: NextRequest) {
   }
 
   // If not authenticated and not on signin page, redirect to signin
-  if (!token && !pathname.startsWith("/signin")) {
+  if (!isAuthenticated && !pathname.startsWith("/signin")) {
     console.log("Redirecting to signin page because user is not authenticated");
     return NextResponse.redirect(new URL("/signin", request.url));
   }
 
-  // If authenticated and on signin page, redirect to dashboard
-  if (token && pathname.startsWith("/signin")) {
+  // If authenticated and on signin page, redirect to incidents
+  if (isAuthenticated && pathname.startsWith("/signin")) {
     console.log(
       "Redirecting to incidents because user try to get /signin but already authenticated"
     );
@@ -47,7 +47,7 @@ export async function middleware(request: NextRequest) {
   }
 
   // Role-based routing (NOC users)
-  if (token?.role === "noc" && !pathname.startsWith("/alerts")) {
+  if (role === "noc" && !pathname.startsWith("/alerts")) {
     return NextResponse.redirect(new URL("/alerts/feed", request.url));
   }
